refactor: :art: consolidate imports and reorganize module structure
