+++ src/pymxbi/__init__.py
@@ -1,4 +1,3 @@
-from .mxbi import MXBI, get_mxbi, set_mxbi
-from .utils import MXBIModel, build_mxbi
+from .mxbi import MXBI, MXBIModel, build_mxbi, get_mxbi, set_mxbi
 
 __all__ = ["MXBI", "MXBIModel", "get_mxbi", "set_mxbi", "build_mxbi"]
